feat: re-organise discord message and documentations
