+++ src/cosmos/tx/actor.rs
@@ -1,6 +1,7 @@
+use crate::cosmos::tx::discord_message::TransactionDiscordMessage;
 use crate::cosmos::tx::messages::trigger::TriggerTx;
 use crate::cosmos::tx::TxHandler;
-use crate::discord::discord_client::message::{DiscordMessage, TransactionMessage};
+use crate::discord::discord_client::message::DiscordMessage;
 use actix::{Actor, AsyncContext, Context};
 use cosmrs::tx::Msg;
 use tracing::info;
@@ -8,7 +9,7 @@ use tracing::info;
 impl<T, M> Actor for TxHandler<T, M>
 where
     T: Msg + Unpin + 'static,
-    M: TransactionMessage + DiscordMessage + Unpin + Send + 'static,
+    M: TransactionDiscordMessage + DiscordMessage + Unpin + Send + 'static,
 {
     type Context = Context<Self>;
