+++ src/cosmos/tx/handlers/register.rs
@@ -1,15 +1,16 @@
 //! Register transaction handler
 
+use crate::cosmos::tx::discord_message::TransactionDiscordMessage;
 use crate::cosmos::tx::messages::register::{RegisterMsg, RegisterMsgResult};
 use crate::cosmos::tx::TxHandler;
-use crate::discord::discord_client::message::{DiscordMessage, TransactionMessage};
+use crate::discord::discord_client::message::DiscordMessage;
 use actix::Handler;
 use cosmrs::tx::Msg;
 
 impl<T, M> Handler<RegisterMsg<T>> for TxHandler<T, M>
 where
     T: Msg + Unpin + 'static,
-    M: TransactionMessage + DiscordMessage + Unpin + Send + 'static,
+    M: TransactionDiscordMessage + DiscordMessage + Unpin + Send + 'static,
 {
     type Result = RegisterMsgResult;
